Simplify props of new card button
